+++ notification-domain/src/main/kotlin/io/github/v1servicenotification/error/NotificationException.kt
@@ -1,0 +1,9 @@
+package io.github.v1servicenotification.error
+
+abstract class NotificationException(
+    val errorCode: ErrorCode
+) : RuntimeException() {
+    override fun fillInStackTrace(): Throwable {
+        return this
+    }
+}
