--- notification-infrastructure/src/main/kotlin/io/github/v1servicenotification/global/error/exception/NotificationException.kt
@@ -1,5 +1,0 @@
-package io.github.v1servicenotification.global.error.exception
-
-open class NotificationException(
-    val errorCode: ErrorCode
-) : RuntimeException()
